+++ backend/tests/integration/test_resilience_endpoints.py
