--- backend/tests/security/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for security utilities."""
